pyproject.toml
[tool.poetry]
name = "vt"
version = "0.1.0"
description = ""
authors = ["Bob Lin <bob401710@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
yt-dlp = "^2023.11.16"
pydantic = "^2.5.2"
click = {version = "^8.1.7"}
ruff = "^0.1.7"
mypy = "^1.7.1"
openai-whisper = {git = "https://github.com/169/whisper.git", rev = "marker"}
loguru = "^0.7.2"
streamlit = "^1.29.0"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vt = "main:cli"